Extract an abstract `graph` package from NVT Workflows (#1265)
* Extract abstract `Node` class from `WorkflowNode`

* Rename `ops.internal` sub-package to `ops.graph`

* Extract `BaseOperator` class from (Workflow) `Operator`

* Move `ColumnSelector` and `Schema` classes into `graph` package

* fix BaseOperator in Node

* Fix `graph.` imports in Node

* Change parent class of graph operators to `BaseOperator`

* Move graph ops into `graph` package

* Remove `WorkflowNode` dependency from `ColumnSelector`

* Fix LambdaOp tests, assorted imports, ColumnSelector rshift

* Move tags into the `graph` package

* Fix additional LambdaOp tests

* Fix doc files paths for `graph` package

* More docs fixes

* Fix a few lambdas that weren't wrapped with `LambdaOp`

* Fix some more naked lambdas

* Move the schema writer into the `graph` package

* Fix tags imports

* Replace fix in `Workflow.input_columns`

* Fix schema writer import

* Remove unused tags import
