+++ docs/source/api/index.rst
@@ -8,7 +8,7 @@ API Documentation
    WorkflowNode <workflow/workflow_node>
    Dataset <dataset>
    Operators <ops/index>
-   ColumnSelector <columns/selector>
-   ColumnSchema <columns/schema>
+   ColumnSelector <graph/selector>
+   ColumnSchema <graph/schema>
    Torch Dataloader <torch_dataloader>
    TensorFlow Dataloader <tensorflow_dataloader>
