+++ nvtabular/graph/ops/identity.py
@@ -14,13 +14,12 @@
 # limitations under the License.
 #
 
-from nvtabular.columns import ColumnSelector
 from nvtabular.dispatch import DataFrameType
+from nvtabular.graph.base_operator import BaseOperator
+from nvtabular.graph.selector import ColumnSelector
 
-from ..operator import Operator
 
-
-class Identity(Operator):
+class Identity(BaseOperator):
     def transform(self, col_selector: ColumnSelector, df: DataFrameType) -> DataFrameType:
         """Simply returns the selected output columns from the input dataframe
