+++ nvtabular/graph/ops/selection.py
@@ -15,13 +15,13 @@
 #
 from __future__ import annotations
 
-from nvtabular.columns import ColumnSelector
-from nvtabular.columns.schema import Schema
 from nvtabular.dispatch import DataFrameType
-from nvtabular.ops import Operator
+from nvtabular.graph.base_operator import BaseOperator
+from nvtabular.graph.schema import Schema
+from nvtabular.graph.selector import ColumnSelector
 
 
-class SelectionOp(Operator):
+class SelectionOp(BaseOperator):
     def __init__(self, selector=None):
         self.selector = selector
