+++ nvtabular/ops/categorify.py
@@ -36,10 +36,10 @@ from pyarrow import parquet as pq
 
 from nvtabular import dispatch
 from nvtabular.dispatch import DataFrameType, _is_cpu_object, _nullable_series, annotate
+from nvtabular.graph.tags import Tags
 from nvtabular.utils import run_on_worker
 from nvtabular.worker import fetch_table_data, get_worker_cache
 
-from ..tags import Tags
 from .operator import ColumnSelector, Operator
 from .stat_operator import StatOperator
