+++ nvtabular/ops/normalize.py
@@ -16,6 +16,8 @@
 import dask.dataframe as dd
 import numpy
 
+from nvtabular.graph.tags import Tags
+
 from ..dispatch import (
     DataFrameType,
     _encode_list_column,
@@ -23,7 +25,6 @@ from ..dispatch import (
     _is_list_dtype,
     annotate,
 )
-from ..tags import Tags
 from .moments import _custom_moments
 from .operator import ColumnSelector, Operator, Supports
 from .stat_operator import StatOperator
