+++ nvtabular/workflow/workflow.py
@@ -32,13 +32,14 @@ import pandas as pd
 from dask.core import flatten
 
 import nvtabular
-from nvtabular.columns import Schema
 from nvtabular.dispatch import _concat_columns
+from nvtabular.graph.node import iter_nodes
+from nvtabular.graph.schema import Schema
 from nvtabular.io.dataset import Dataset
 from nvtabular.ops import StatOperator
 from nvtabular.utils import _ensure_optimize_dataframe_graph, global_dask_client
 from nvtabular.worker import clean_worker_cache
-from nvtabular.workflow.node import WorkflowNode, iter_nodes
+from nvtabular.workflow.node import WorkflowNode
 
 LOG = logging.getLogger("nvtabular")
