+++ tests/integration/test_nvt_hugectr.py
@@ -88,7 +88,9 @@ def test_training():
 
     # Perform ETL with NVTabular
     cat_features = CATEGORICAL_COLUMNS >> nvt.ops.Categorify(cat_cache="device")
-    ratings = nvt.ColumnSelector(["rating"]) >> (lambda col: (col > 3).astype("int8"))
+    ratings = nvt.ColumnSelector(["rating"]) >> nvt.ops.LambdaOp(
+        lambda col: (col > 3).astype("int8")
+    )
     output = cat_features + ratings
 
     workflow = nvt.Workflow(output)
