+++ tests/unit/loader/test_tf_dataloader.py
@@ -566,7 +566,9 @@ def test_horovod_multigpu(tmpdir):
     df_files = df_gen.full_df_create(10000, cols, output=target_path)
     # process them
     cat_features = nvt.ColumnSelector(["userId", "movieId", "genres"]) >> nvt.ops.Categorify()
-    ratings = nvt.ColumnSelector(["rating"]) >> (lambda col: (col > 3).astype("int8"))
+    ratings = nvt.ColumnSelector(["rating"]) >> nvt.ops.LambdaOp(
+        lambda col: (col > 3).astype("int8")
+    )
     output = cat_features + ratings
     proc = nvt.Workflow(output)
     target_path_train = os.path.join(tmpdir, "train/")
